+++ lib/salve/parse.ts
@@ -11,8 +11,7 @@ import * as fs from "fs";
 import * as path from "path";
 import * as sax from "sax";
 
-import { convertRNGToPattern, Event, Grammar,
-         readTreeFromJSON } from "./validate";
+import { convertRNGToPattern, Grammar, readTreeFromJSON } from "./validate";
 
 // tslint:disable no-console
 
@@ -84,14 +83,13 @@ export async function parse(rngSource: string | Grammar,
 
   let error = false;
 
-  function fireEvent(...args: any[]): void {
-    const ev = new Event(args);
-    const ret = walker.fireEvent(ev);
+  function fireEvent(name: string, args: any[]): void {
+    const ret = walker.fireEvent(name, args);
     if (ret instanceof Array) {
       error = true;
       if (!mute) {
         for (const err of ret) {
-          console.log(`on event ${ev}`);
+          console.log(`on event ${name}, ${args.join(", ")}`);
           console.log(err.toString());
         }
       }
@@ -103,7 +101,7 @@ export async function parse(rngSource: string | Grammar,
 
   function flushTextBuf(): void {
     if (textBuf !== "") {
-      fireEvent("text", textBuf);
+      fireEvent("text", [textBuf]);
       textBuf = "";
     }
   }
@@ -128,16 +126,16 @@ export async function parse(rngSource: string | Grammar,
       }
     }
     if (nsDefinitions.length !== 0) {
-      fireEvent("enterContext");
+      fireEvent("enterContext", []);
       for (const definition of nsDefinitions) {
-        fireEvent("definePrefix", ...definition);
+        fireEvent("definePrefix", definition);
       }
     }
-    fireEvent("enterStartTag", node.uri, node.local);
+    fireEvent("enterStartTag", [node.uri, node.local]);
     for (const event of attributeEvents) {
-      fireEvent(...event);
+      fireEvent(event[0], event.slice(1));
     }
-    fireEvent("leaveStartTag");
+    fireEvent("leaveStartTag", []);
     tagStack.unshift({
       uri: node.uri,
       local: node.local,
@@ -155,9 +153,9 @@ export async function parse(rngSource: string | Grammar,
     if (tagInfo === undefined) {
       throw new Error("stack underflow");
     }
-    fireEvent("endTag", tagInfo.uri, tagInfo.local);
+    fireEvent("endTag", [tagInfo.uri, tagInfo.local]);
     if (tagInfo.hasContext) {
-      fireEvent("leaveContext");
+      fireEvent("leaveContext", []);
     }
   };