+++ test/validation.js
@@ -67,7 +67,7 @@ function makeParser(er, walker) {
       er.recordEvent(walker, "attributeName", uri, local);
       er.recordEvent(walker, "attributeValue", value);
     }
-    er.recordEvent(walker, "leaveStartTag", node.uri, node.local);
+    er.recordEvent(walker, "leaveStartTag", []);
     tagStack.unshift([node.uri, node.local]);
   };
 
@@ -123,14 +123,13 @@ class EventRecorder {
       this.recorded_states.push([walker.clone(), this.ce.exp_ix, evIx]);
     }
 
-    const nev = new salve.Event(evParams);
     if (this.check_fireEvent_invocation) {
       this.ce.compare(
-        `\ninvoking fireEvent with ${nev.toString().trim()
-.replace(/\s+\n/g, "\n")}`, nev);
+        `\ninvoking fireEvent with Event: ${evParams.join(", ").trim()
+.replace(/\s+\n/g, "\n")}`, evParams);
     }
-    const ret = walker.fireEvent(nev);
-    this.ce.compare(`fireEvent returned ${errorsToString(ret)}`, nev);
+    const ret = walker.fireEvent(evParams[0], evParams.slice(1));
+    this.ce.compare(`fireEvent returned ${errorsToString(ret)}`, evParams);
     if (this.check_possible) {
       const possibleEvs = Array.from(walker.possible());
       // We sort events alphabetically, because the
@@ -140,7 +139,7 @@ class EventRecorder {
           evParams[0] !== "leaveContext" &&
           evParams[0] !== "definePrefix") {
         this.ce.compare(
-          `possible events\n${salve.eventsToTreeString(possibleEvs)}`, nev);
+          `possible events\n${salve.eventsToTreeString(possibleEvs)}`, evParams);
       }
     }
   }
@@ -163,7 +162,7 @@ class ComparisonEngine {
     const to = this.expected.slice(this.exp_ix, this.exp_ix + lines.length);
 
     assert.equal(msg, to.join("\n"),
-                 `at line: ${this.exp_ix + 1} event ${ev.toString()}`);
+                 `at line: ${this.exp_ix + 1} event ${ev.join(", ")}`);
     this.exp_ix += lines.length;
   }
 }
@@ -196,15 +195,15 @@ function makeValidTest(dir) {
 
     const contextIndependent = tree.whollyContextIndependent();
     ce.compare(`wholly context-independent ${contextIndependent}`,
-               "*context-independent*");
+               ["*context-independent*"]);
 
     ce.compare(`possible events\n${salve.eventsToTreeString(walker.possible())}`,
-               new salve.Event(["initial"]));
+               ["initial"]);
 
     const parser = makeParser(er, walker);
     parser.write(xmlSource).close();
 
-    ce.compare(`end returned ${walker.end()}`, "*final*");
+    ce.compare(`end returned ${walker.end()}`, ["*final*"]);
 
     // Roll back; >> gives us an integer
     // eslint-disable-next-line no-bitwise
@@ -218,7 +217,7 @@ function makeValidTest(dir) {
       more = er.issueEventAt(walker, evIx++);
     }
 
-    ce.compare(`end returned ${walker.end()}`, "*final*");
+    ce.compare(`end returned ${walker.end()}`, ["*final*"]);
   };
 }
 
@@ -304,7 +303,7 @@ describe("GrammarWalker.fireEvent", () => {
 
         const parser = makeParser(er, walker);
         parser.write(xmlSource).close();
-        ce.compare(`end returned ${walker.end()}`, "*final*");
+        ce.compare(`end returned ${walker.end()}`, ["*final*"]);
       };
     }
 
@@ -414,11 +413,11 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "style"));
+      ret = walker.fireEvent("attributeName", ["", "style"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "style"));
+      ret = walker.fireEvent("attributeName", ["", "style"]);
       assert.equal(ret.length, 1);
       assert.equal(
         ret[0].toString(),
@@ -434,9 +433,9 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag", "", "html"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
 
       const permutations = [
@@ -451,7 +450,7 @@ describe("GrammarWalker.fireEvent", () => {
               "attributeName:\n" +
               "    ";
       for (const perm of permutations) {
-        ret = walker.fireEvent(new salve.Event("enterStartTag", "", "em"));
+        ret = walker.fireEvent("enterStartTag", ["", "em"]);
         assert.isFalse(ret, "entering em");
 
         const possible = [];
@@ -463,14 +462,14 @@ describe("GrammarWalker.fireEvent", () => {
           assert.equal(salve.eventsToTreeString(walker.possible()),
                        `${stub}${sorted.join("\n    ")}\n`);
 
-          ret = walker.fireEvent(new salve.Event("attributeName", "", attr));
+          ret = walker.fireEvent("attributeName", ["", attr]);
           assert.isFalse(ret);
 
           // We've seen it. This array is in the same order
           // as perm.
           possible.shift();
 
-          ret = walker.fireEvent(new salve.Event("attributeValue", "x"));
+          ret = walker.fireEvent("attributeValue", ["x"]);
           assert.isFalse(ret);
 
           // Seen all possible attributes.
@@ -480,9 +479,9 @@ describe("GrammarWalker.fireEvent", () => {
           }
         }
 
-        ret = walker.fireEvent(new salve.Event("leaveStartTag"));
+        ret = walker.fireEvent("leaveStartTag", []);
         assert.isFalse(ret);
-        ret = walker.fireEvent(new salve.Event("endTag", "", "em"));
+        ret = walker.fireEvent("endTag", ["", "em"]);
         assert.isFalse(ret);
       }
     });
@@ -494,20 +493,20 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag", "", "html"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
 
-      ret = walker.fireEvent(new salve.Event("enterStartTag", "", "em"));
+      ret = walker.fireEvent("enterStartTag", ["", "em"]);
       assert.isFalse(ret, "entering em");
-      ret = walker.fireEvent(new salve.Event("leaveStartTag"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.deepEqual(ret.map(x => x.toString()), [
         "attribute missing: {\"ns\":\"\",\"name\":\"attr-a\"}",
         "attribute missing: {\"ns\":\"\",\"name\":\"attr-b\"}",
         "attribute missing: {\"ns\":\"\",\"name\":\"attr-c\"}",
       ]);
-      ret = walker.fireEvent(new salve.Event("endTag", "", "em"));
+      ret = walker.fireEvent("endTag", ["", "em"]);
       assert.deepEqual(ret.map(x => x.toString()), [
         "tag required: {\"ns\":\"\",\"name\":\"foo\"}",
       ]);
@@ -523,35 +522,35 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag", "", "html"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
 
-      ret = walker.fireEvent(new salve.Event("enterStartTag", "", "em"));
+      ret = walker.fireEvent("enterStartTag", ["", "em"]);
       assert.isFalse(ret, "entering em");
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "attr-a"));
+      ret = walker.fireEvent("attributeName", ["", "attr-a"]);
       assert.isFalse(ret, "attr-a");
-      ret = walker.fireEvent(new salve.Event("attributeValue", "x"));
+      ret = walker.fireEvent("attributeValue", ["x"]);
       assert.isFalse(ret, "attr-a value");
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "attr-b"));
+      ret = walker.fireEvent("attributeName", ["", "attr-b"]);
       assert.isFalse(ret, "attr-b");
-      ret = walker.fireEvent(new salve.Event("attributeValue", "x"));
+      ret = walker.fireEvent("attributeValue", ["x"]);
       assert.isFalse(ret, "attr-b value");
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "attr-c"));
+      ret = walker.fireEvent("attributeName", ["", "attr-c"]);
       assert.isFalse(ret, "attr-c");
-      ret = walker.fireEvent(new salve.Event("attributeValue", "x"));
+      ret = walker.fireEvent("attributeValue", ["x"]);
       assert.isFalse(ret, "attr-c value");
-      ret = walker.fireEvent(new salve.Event("leaveStartTag"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret, "leaveStartTag has no errors");
 
-      ret = walker.fireEvent(new salve.Event("enterStartTag", "", "foo"));
+      ret = walker.fireEvent("enterStartTag", ["", "foo"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag", "", "foo"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("endTag", "", "foo"));
+      ret = walker.fireEvent("endTag", ["", "foo"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("endTag", "", "em"));
+      ret = walker.fireEvent("endTag", ["", "em"]);
       assert.deepEqual(ret.map(x => x.toString()), [
         "must choose either {\"ns\":\"\",\"name\":\"bar\"} or " +
           "{\"ns\":\"\",\"name\":\"baz\"}",
@@ -568,27 +567,27 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag", "", "html"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
 
-      ret = walker.fireEvent(new salve.Event("enterStartTag", "", "em"));
+      ret = walker.fireEvent("enterStartTag", ["", "em"]);
       assert.isFalse(ret, "entering em");
-      ret = walker.fireEvent(new salve.Event("leaveStartTag"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.deepEqual(ret.map(x => x.toString()), [
         "attribute missing: {\"ns\":\"\",\"name\":\"attr-a\"}",
         "attribute missing: {\"ns\":\"\",\"name\":\"attr-b\"}",
         "attribute missing: {\"ns\":\"\",\"name\":\"attr-c\"}",
       ]);
 
-      ret = walker.fireEvent(new salve.Event("enterStartTag", "", "foo"));
+      ret = walker.fireEvent("enterStartTag", ["", "foo"]);
       assert.isFalse(ret);
-      // ret = walker.fireEvent(new salve.Event("leaveStartTag", "", "foo"));
+      // ret = walker.fireEvent("leaveStartTag", []);
       // assert.isFalse(ret);
-      // ret = walker.fireEvent(new salve.Event("endTag", "", "foo"));
+      // ret = walker.fireEvent("endTag", ["", "foo"]);
       // assert.isFalse(ret);
-      // ret = walker.fireEvent(new salve.Event("endTag", "", "em"));
+      // ret = walker.fireEvent("endTag", ["", "em"]);
     });
   });
 
@@ -605,9 +604,9 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("text", "q"));
+      ret = walker.fireEvent("text", ["q"]);
       assert.equal(ret.length, 1);
       assert.equal(ret[0].toString(), "text not allowed here");
     });
@@ -618,15 +617,15 @@ describe("GrammarWalker.fireEvent", () => {
 
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
-      let ret = walker.fireEvent(new salve.Event("enterStartTag", "", "html"));
+      let ret = walker.fireEvent("enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "style"));
+      ret = walker.fireEvent("attributeName", ["", "style"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeValue", "", "x"));
+      ret = walker.fireEvent("attributeValue", ["", "x"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
-      assert.throws(() => walker.fireEvent(new salve.Event("leaveStartTag")),
+      assert.throws(() => walker.fireEvent("leaveStartTag", []),
                     Error,
                     "unexpected leaveStartTag event; it is likely that \
 fireEvent is incorrectly called");
@@ -639,13 +638,13 @@ fireEvent is incorrectly called");
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
       let ret = walker.fireEvent(
-        new salve.Event("enterStartTag", "", "html"));
+        "enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "style"));
+      ret = walker.fireEvent("attributeName", ["", "style"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeValue", "", "x"));
+      ret = walker.fireEvent("attributeValue", ["", "x"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeValue", "", "x"));
+      ret = walker.fireEvent("attributeValue", ["", "x"]);
       assert.equal(ret.length, 1);
       assert.equal(ret[0].toString(),
                    "unexpected attributeValue event; " +
@@ -660,19 +659,19 @@ fireEvent is incorrectly called");
       const tree = salve.readTreeFromJSON(source);
       const walker = tree.newWalker();
       let ret = walker.fireEvent(
-        new salve.Event("enterStartTag", "", "html"));
+        "enterStartTag", ["", "html"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeName", "", "style"));
+      ret = walker.fireEvent("attributeName", ["", "style"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("attributeValue", "", "x"));
+      ret = walker.fireEvent("attributeValue", ["", "x"]);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("leaveStartTag"));
+      ret = walker.fireEvent("leaveStartTag", []);
       assert.isFalse(ret);
-      ret = walker.fireEvent(new salve.Event("endTag", "", "html"));
+      ret = walker.fireEvent("endTag", ["", "html"]);
       assert.equal(ret.length, 1);
       assert.equal(ret[0].toString(),
                    "tag required: {\"ns\":\"\",\"name\":\"head\"}");
-      ret = walker.fireEvent(new salve.Event("endTag", "", "html"));
+      ret = walker.fireEvent("endTag", ["", "html"]);
       assert.equal(ret.length, 1);
       assert.equal(ret[0].toString(),
                    "unexpected end tag: {\"ns\":\"\",\"name\":\"html\"}");